requirements.txt
streamlit==1.32.0
pillow==10.2.0
plotly==5.18.0 